+++ packages/agents-ui/src/hooks/index.ts
@@ -5,3 +5,5 @@ export { useHandler } from "./useHandler";
 export { useFilteredEvents } from "./useFilteredEvents";
 export { useTimeAgo } from "./useTimeAgo";
 export { useRouterContext } from "./useRouterContext";
+export { useAgentMCP } from "./useAgentMCP";
+export { useCopyToClipboard } from "./useCopyToClipboard";